FileMetadataProvider: Use custom translations for some properties
In the KFileMetadata widget the tags, rating, and comment were given as
a custom property (kfileitem#tags). We now use the standard SDO
properties instead. However, we do not want the translation string to
show "has tag" and "numeric rating". We want the old strings of "Tags",
"Ratings" and "Comment". Hence this hack.