FileMetadataProvider: Improve data merging when provided with multiple resources
For all the resources, each property is compared with the other
resources and the least common denomiator is shown to the user.

Eg - If the album is common to all the resources, it will be shown. For
resources which have been tagged, only the tags which are present in all
the resources will be shown.