WidgetFactory: Properly handle multiple file tagging
When multiple files have been selected then setProperty should not be
used, as the tag widget will only show the common tags, and changing
that list of common tags will result in the not-common tags being
removed.

Instead, we use addProperty to add the changed tags. This way the old
tags are preserved. Additionally, we maintain a list of the previous
tags, and calculate the difference to remove the previous tags.

BUG: 160157
FIXED-IN: 4.11