MetadataFilter: Do not show editable data for contacts and albums
Though it does make sense for the user to be able to tag/rate contacts
and albums, dolphin is not the place to do it from. Specially since
contacts and albums are not displayed by the nepomuk kioslaves.

That would result in the user tagging the contact/album and it then not
appearing in the tag results.