FileMetaDataProvider: Do not show editable data in Nepomuk is not running
Editable Data = Tags, Ratings and comments

Normally, the editable data is not shown if Nepomuk is not running.
However, the check for nepomuk is only done once during startup, this
results in the editable data being shown if the widget is still running
and nepomuk is switched off.

With this patch, we do the check each time.