FileMetadataProvider: Index the file if it has indexingLevel = 1
With 4.10 the file indexer is only called when the user is idle. This
means that the user could select a file and its information will not be
shown in the sidebar unless it has been indexed. This is not ideal. The
users want to be able to see the indexed information. They don't want to
have to wait for the indexing to complete.

Now, whenever the FileMetadataWidget displays the results of a file
which has not been completely indexed, it calls the nepmukindexer
process to index it completely.