FileMetadataProvider: Send the data in multiple chunks
When more than one file item has been selected we should send the data
in multiple chunks. Inititally we just send the basic data (total file
size), and later when the ResourceData has been loaded we send all of
it.

This results in the FileMetadataWidget at least showing something for
very large selections.