ResourceLoader: Load all the associated properties as well
When using the FileMetadataWidget, the MetadatFilter class is used which
internally uses Types::Property. This Property class is unfortunately
synchronous and can block for a bit. Therefore we load all the
properties in this separate thread so that they are always there in the
cache.

This never used to be a problem earlier because connecting to the
Nepomuk Server used to be quite fast. Since now we connect directly to
virtuoso, it can take some time and block.

BUG: 321838