Port tagging infrastructure to Baloo
Removed a lot of the additional functionality provided by the
TagWidget. It is now just a simple widget to view tags and to launch the
KEditTagsDialog. It no longer has that additional resource handling
code.

This way we only target our main use case which is the
FileMetaDataWidget and ignore the alternative use case which was not
really used.