[tagsaction] Don't crash on multiple files
Summary:
Tagging multiple files is not supported and thus we never create m_metadata for it. However, we start the tags listing  and as soon as tags are found m_metadata is accesses which results in a crash.

This patch delays the tags listing until m_metadata has been initialized

BUG: 401733

Test Plan: Select multiple files in Dolphin, right click -> no crash

Reviewers: elvisangelaccio

Reviewed By: elvisangelaccio

Subscribers: #baloo

Tags: #baloo

Differential Revision: https://phabricator.kde.org/D20361