Make the FileMetadataWidget stretch vertically in a sensible way
Summary:
When the widget size is larger than the required size, the QGridLayout
distributes the extra space evenly between rows, leading awkward blank
areas between rows.
Though this can be avoided by wrapping the widget into e.g. a VBoxLayout
with a spacer at the bottom it leads to unnecessarily complicated
layouts where the widget is used.
It is much simpler to just insert a spacer as the last row in the
QGridLayout.

Depends on D20524

Reviewers: #dolphin, #baloo, #frameworks, ngraham, astippich

Reviewed By: #dolphin, #baloo, ngraham

Tags: #baloo

Differential Revision: https://phabricator.kde.org/D20667