[UserMetaData] Consolidate UserMetaData conversion to common implementation
Summary:
The conversion from UserMetaData tags/rating etc. was implemented in both
the standalone temporary extractor and for the baloo DB cached code path,
although the standalone implementation lacked support for the originUrl
entry.

Consolidate the implementations to a common helper, and move the call
from the standalone extractor process to its calling job. This avoids
the need to serialize the data, doing it from the job context avoids
blocking the UI.

Test Plan: ctest

Reviewers: #baloo, astippich, ngraham

Reviewed By: astippich

Tags: #baloo

Differential Revision: https://phabricator.kde.org/D20956