[FileMetaDataProvider] Move readOnly check into the fetch jobs
Summary:
UserMetaData::isSupported() is blocking, so it should be done
asynchronously. Also, the jobs already instantiate a UserMetaData for
each file, so doing it twice is wasteful.

Instead of doing the check in the widget class and pushing it to the
provider, let the provider determine if the file attributes can be
modified and forward the result to the widget factory after the provider
has finished.

Do not check if the file is local twice, once in the provider is
sufficient.

Remove one bogus test case.

Depends on D20979

Test Plan: ctest

Reviewers: #baloo, astippich, ngraham

Reviewed By: astippich

Tags: #baloo

Differential Revision: https://phabricator.kde.org/D20980