Cleanup default filters for baloo-widgets
Summary:
The property names do not match any names defined in KFileMetaData.
While the intention of the previous filters clearly was to hide most of the EXIF information,
users seem to use the provided information (see bug report), so show most of it.
Hide photo metering mode, photo sharpness and white balance, as they seem to be quite useless.

CCBUG: 343273

Reviewers: ngraham, bruns

Reviewed By: bruns

Subscribers: #baloo

Tags: #baloo

Differential Revision: https://phabricator.kde.org/D21413