Do not display a buggy filesize for non local files
Summary:
When displaying a virtual folder or remote drive the folder size in the information panel is wrong.
The subDirectoriesCount function supports only local files.
It creates a bug where when a url such as recentdocuments:/ is displayed in dolphin, the count reflected in the size field is in fact the count of its path meaning the number of files in the root folder /.

Test Plan: Manual

Reviewers: #dolphin, ngraham, elvisangelaccio

Reviewed By: #dolphin, elvisangelaccio

Subscribers: #baloo

Tags: #baloo

Differential Revision: https://phabricator.kde.org/D22529