Split handling of on-demand extractor process from IndexedDataRetriever
Summary:
To allow a unification of the IndexedDataRetriever job and FileFetchJob,
move the process handling into a separate class and only keep
the UserMetaData handling.

As a small optimization, interleave the forking of the external process,
UserMetaData retrieval and joining of the external process. Also reuse
the QFileInfo in the constructor to determine if the file is writable and
thus has editable metadata.

Beyond the optimization, no user visible changes.

Depends on D25247

Test Plan:
When using a location not indexed by baloo, metadata is
still shown as before

Reviewers: #baloo, ngraham, astippich

Reviewed By: #baloo, ngraham, astippich

Tags: #baloo

Differential Revision: https://phabricator.kde.org/D25248