TagsFileItemAction: Make sure we always create a menu
We need to make sure the m_menu is always created
when actions are fetched,
since if m_menu would get deleted every time the context menu is
closed, this would cause the m_menu become a dangling pointer and crash when context
menu is opened the second time.

Instead, just create a new menu and set its parent every time.
We also need to set the newTag parent to menus parent, so that
the popup wont immediately be deleted when the context menu this
action is in is closed.